requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.1
cycler==0.10.0
dask==1.1.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-mysql==2.4.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
imutils==0.5.2
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
networkx==2.2
numpy==1.15.4
oauthlib==3.0.0
opencv-contrib-python==3.4.5.20
Pillow==5.4.1
psycopg2==2.7.7
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-image==0.14.2
scipy==1.2.0
six==1.12.0
toolz==0.9.0
urllib3==1.24.1
virtualenv==16.2.0
whitenoise==4.1.2
wrapt==1.11.1
